+++ bundle/src/test/java/com/adobe/acs/commons/data/VariantTest.java
@@ -22,6 +22,8 @@ package com.adobe.acs.commons.data;
 import java.text.SimpleDateFormat;
 import java.time.Instant;
 import java.util.Date;
+import java.util.Locale;
+
 import org.junit.Test;
 
 import static org.junit.Assert.*;
@@ -68,8 +70,11 @@ public class VariantTest {
         assertEquals(nowInstant, Variant.convert(now, Instant.class));
         assertEquals(nowInstant, Variant.convert(nowDate, Instant.class));
 
-        String nowStringShort = SimpleDateFormat.getDateTimeInstance(SimpleDateFormat.SHORT, SimpleDateFormat.LONG).format(nowDate);
-        String nowStringLong = SimpleDateFormat.getDateTimeInstance(SimpleDateFormat.LONG, SimpleDateFormat.LONG).format(nowDate);
+        // Locale.getDefault() and Locale.getDefault(Locale.Category.FORMAT) may return different values in certain OS settings.
+        // Variant uses the former, SimpleDateFormat uses the latter by default.
+        // To make things consistent, pass Locale.getDefault() to SimpleDateFormat explicitly.
+        String nowStringShort = SimpleDateFormat.getDateTimeInstance(SimpleDateFormat.SHORT, SimpleDateFormat.LONG, Locale.getDefault()).format(nowDate);
+        String nowStringLong = SimpleDateFormat.getDateTimeInstance(SimpleDateFormat.LONG, SimpleDateFormat.LONG, Locale.getDefault()).format(nowDate);
         assertNotNull(Variant.convert(nowStringLong, Date.class).getTime());
         assertNotNull(Variant.convert(nowStringShort, Date.class).getTime());
         assertNotNull(Variant.convert("12:00 AM", Date.class).getTime());